requirements.txt
PyInstaller==2.1
PyYAML==3.11
argparse==1.2.1
configobj==5.0.4
paramiko
pywinrm
ipcalc
validate
zxcvbn
pyasn1
